OntarioOneCallChallenge/Setup/requirements.txt
et-xmlfile==1.1.0
mysql-connector-python==8.0.30
numpy==1.23.2
openpyxl==3.0.10
pandas==1.4.3
protobuf==3.20.1
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
